refactor: 기도 제목/내용 편집 UI 개선 및 불필요한 update 화면 제거 #47
- PrayerDetailScreen을 PrayerReadBody로 이름 변경하여 일관성 개선
- 기도 제목/내용 수정 시 인라인 편집 대신 dialog 사용으로 UX 통일
 - 편집 모드 토글을 헤더 버튼으로 이동 (Switch 컴포넌트 제거)
 - 수정/삭제 시 확인 팝업 추가 및 뒤로가기 버튼 처리 개선
 - dialog 위치를 화면 상단으로 조정 (paddingTop: 50)
 - 긴 텍스트 수정 시 스크롤 위치를 맨 위로 자동 설정
 - 불필요한 update 관련 컴포넌트 및 store 완전 제거:
  * /update 폴더 및 하위 컴포넌트들
  * usePrayerUpdateStore, PrayerUpdateItem 등 관련 타입
  * updatePrayerParams, updatePrayerRequest 등 미사용 타입
 - EditableTitleCard 컴포넌트를 TitleCard + dialog 조합으로 대체
