+++ src/common/components/modal/ConfirmationModal.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import { StyleSheet, View, Text, Dimensions } from "react-native";
+import React, { useEffect } from "react";
+import { StyleSheet, View, Text, Dimensions, BackHandler } from "react-native";
 import { Button, Portal, Modal, IconButton } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
 import { color } from "@/common/styles/color";
@@ -29,6 +29,21 @@ function ConfirmationModal({
   cancelText = "취소",
   iconColor = color.secondary,
 }: ConfirmationModalProps) {
+  // 백 버튼 핸들러 추가 - 일단 주석 처리
+  // useEffect(() => {
+  //   if (visible) {
+  //     const backHandler = BackHandler.addEventListener(
+  //       'hardwareBackPress',
+  //       () => {
+  //         onDismiss();
+  //         return true; // 기본 뒤로가기 동작 막기
+  //       }
+  //     );
+
+  //     return () => backHandler.remove();
+  //   }
+  // }, [visible, onDismiss]);
+
   return (
     <Portal>
       <Modal
