+++ src/domain/prayers/services/prayerServices.ts
@@ -6,8 +6,6 @@ import { MessageResponse } from "../../../common/types/messageResponse";
 import { PrayerTitle } from "../types/prayerTitle";
 import { FetchPrayerTitlesResponse } from "../types/response/fetchPrayerTitlesResponse";
 import type { FetchPrayerContentsResponse } from "../types/response/fetchPrayerContentsResponse";
-import { PrayerUpdateItem } from "../types/prayerUpdateItem";
-import { UpdatePrayerRequest } from "../types/request/updatePrayerRequest";
 import { CreatePrayerCompletionRequest } from "../types/request/createPrayerCompletionRequest";
 
 export const prayerService = {
@@ -59,22 +57,6 @@ export const prayerService = {
     return response.data.prayerContents ?? [];
   },
   // 기도 제목 변경
-  update: async (
-    prayerTitleId: number | null,
-    title: string,
-    prayerList: PrayerUpdateItem[]
-  ): Promise<MessageResponse> => {
-    const response: MessageResponse = await apiService.put<UpdatePrayerRequest>(
-      `/prayers/${prayerTitleId}`,
-      {
-        title,
-        contents: prayerList,
-      } as UpdatePrayerRequest
-    );
-    console.log("API response=", response.message);
-    return response;
-  },
-
   // 기도 제목만 수정
   updateTitle: async (
     prayerTitleId: number | null,
