--- src/domain/prayers/types/prayerUpdateItem.ts
@@ -1,3 +1,0 @@
-import type { PrayerContent } from "./prayerContent";
-
-export interface PrayerUpdateItem extends PrayerContent {}
