--- src/domain/prayers/types/request/updatePrayerRequest.ts
@@ -1,6 +1,0 @@
-import { PrayerUpdateItem } from "../prayerUpdateItem";
-
-export interface UpdatePrayerRequest {
-  title: string;
-  contents: PrayerUpdateItem[];
-}
